feat: add Designer app, tracker services, LiveMap improvements, and fix template schema
- Add SocialMotive.Designer Blazor app with canvas editor, SignalR hub, and render worker
- Add DesignController, DesignService, and BackgroundRenderQueue for template rendering
- Fix DbTemplate entity to match DB schema (Name, Description, Width, Height, Tags, Category as direct columns)
- Move template metadata out of TemplateJson — JSON now only stores Layers
- Simplify GeneratorMappingProfile to map fields directly from DbTemplate
- Remove unused migration 012 (template JSON consolidation)
- Add TrackerService, TrackerTimeoutService, and LocationCacheService
- Refactor LiveMap: remove Google Maps, improve LeafletMap, add tracker timeout handling
- Update TelegramBot configuration and handler
- Various model, mapping, and controller improvements across Admin, Generator, and LiveMap areas

+++ src/SocialMotive.Core/Data/DbLocation.cs
@@ -4,12 +4,12 @@ namespace SocialMotive.Core.Data
     {
         public long LocationId { get; set; }
         public int TrackerId { get; set; }
-        public float Latitude { get; set; }
-        public float Longitude { get; set; }
-        public float? AccuracyMeters { get; set; }
-        public float? AltitudeMeters { get; set; }
-        public float? SpeedKmh { get; set; }
-        public float? HeadingDegrees { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double? AccuracyMeters { get; set; }
+        public double? AltitudeMeters { get; set; }
+        public double? SpeedKmh { get; set; }
+        public double? HeadingDegrees { get; set; }
         public DateTime Timestamp { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime ModifiedAt { get; set; }
