+++ src/SocialMotive.Core/Model/LiveMap/TrackerLocation.cs
@@ -4,10 +4,10 @@ public class TrackerLocation
 {
     public int TrackerId { get; set; }
     public string DisplayName { get; set; } = string.Empty;
-    public float Latitude { get; set; }
-    public float Longitude { get; set; }
-    public float? AccuracyMeters { get; set; }
-    public float? SpeedKmh { get; set; }
-    public float? HeadingDegrees { get; set; }
+    public double Latitude { get; set; }
+    public double Longitude { get; set; }
+    public double? AccuracyMeters { get; set; }
+    public double? SpeedKmh { get; set; }
+    public double? HeadingDegrees { get; set; }
     public DateTime Timestamp { get; set; }
 }
